[Refactor] Move query queue to storage module
