+++ be/src/exec/meta_scanner.h
@@ -18,9 +18,9 @@
 
 #include "common/status.h"
 #include "exec/meta_scan_node.h"
-#include "exec/olap_utils.h"
 #include "runtime/runtime_state_fwd.h"
 #include "storage/meta_reader.h"
+#include "storage/query/olap_scan_range.h"
 
 namespace starrocks {
