+++ be/src/exec/olap_common.h
@@ -41,9 +41,9 @@
 #include <utility>
 #include <vector>
 
-#include "exec/olap_utils.h"
 #include "storage/primitive/column_value_range.h"
 #include "storage/primitive/olap_tuple.h"
+#include "storage/query/olap_scan_range.h"
 
 namespace starrocks {
