+++ be/src/exec/olap_scan_node.cpp
@@ -40,11 +40,9 @@
 #include "exec/pipeline/pipeline_builder_operators.h"
 #include "exec/pipeline/scan/chunk_buffer_limiter.h"
 #include "exec/pipeline/scan/morsel_queue_factory.h"
-#include "exec/pipeline/scan/olap_fixed_morsel_queue_builder.h"
 #include "exec/pipeline/scan/olap_scan_operator.h"
 #include "exec/pipeline/scan/olap_scan_prepare_operator.h"
 #include "exec/pipeline/scan/scan_morsel.h"
-#include "exec/pipeline/scan/split_morsel_queue_builder.h"
 #include "exprs/column_access_path_resolver.h"
 #include "exprs/expr.h"
 #include "exprs/expr_context.h"
@@ -59,6 +57,8 @@
 #include "storage/chunk_helper.h"
 #include "storage/primitive/storage_ids.h"
 #include "storage/primitive/storage_version.h"
+#include "storage/query/olap_fixed_morsel_queue_builder.h"
+#include "storage/query/split_morsel_queue_builder.h"
 #include "storage/rowset/rowset.h"
 #include "storage/storage_engine.h"
 #include "storage/tablet.h"
