+++ be/src/exec/pipeline/scan/lake_meta_scan_prepare_operator.cpp
@@ -18,8 +18,8 @@
 
 #include "exec/pipeline/scan/meta_scan_context.h"
 #include "exec/pipeline/scan/meta_scan_operator.h"
-#include "exec/pipeline/scan/olap_morsel_queue.h"
 #include "gen_cpp/Types_types.h"
+#include "storage/query/olap_morsel_queue.h"
 
 namespace starrocks::pipeline {
