+++ be/src/exec/pipeline/scan/morsel.h
@@ -14,11 +14,11 @@
 
 #pragma once
 
-#include "exec/pipeline/scan/bucket_sequence_morsel_queue.h"
 #include "exec/pipeline/scan/morsel_queue.h"
 #include "exec/pipeline/scan/morsel_queue_factory.h"
-#include "exec/pipeline/scan/olap_dynamic_morsel_queue.h"
-#include "exec/pipeline/scan/olap_fixed_morsel_queue.h"
-#include "exec/pipeline/scan/olap_morsel_queue.h"
-#include "exec/pipeline/scan/split_morsel_queue.h"
 #include "exec/pipeline/scan/ticketed_morsel_queue.h"
+#include "storage/query/bucket_sequence_morsel_queue.h"
+#include "storage/query/olap_dynamic_morsel_queue.h"
+#include "storage/query/olap_fixed_morsel_queue.h"
+#include "storage/query/olap_morsel_queue.h"
+#include "storage/query/split_morsel_queue.h"
