+++ be/src/exec/scan_node.cpp
@@ -38,11 +38,11 @@
 #include "exec/pipeline/query_context.h"
 #include "exec/pipeline/scan/morsel_queue_builder.h"
 #include "exec/pipeline/scan/morsel_queue_factory.h"
-#include "exec/pipeline/scan/olap_dynamic_morsel_queue.h"
-#include "exec/pipeline/scan/olap_fixed_morsel_queue_builder.h"
 #include "exec/pipeline/scan/scan_morsel.h"
 #include "exprs/expr_factory.h"
 #include "runtime/runtime_state.h"
+#include "storage/query/olap_dynamic_morsel_queue.h"
+#include "storage/query/olap_fixed_morsel_queue_builder.h"
 
 namespace starrocks {
