+++ be/src/storage/lake/tablet_reader.cpp
@@ -26,9 +26,7 @@
 #include "common/config_scan_io_fwd.h"
 #include "common/status.h"
 #include "common/thread/threadpool.h"
-#include "exec/pipeline/scan/morsel.h"
 #include "exec/pipeline/scan/scan_morsel.h"
-#include "exec/pipeline/scan/split_scan_morsel.h"
 #include "gutil/stl_util.h"
 #include "runtime/env/global_env.h"
 #include "runtime/runtime_state.h"
@@ -45,6 +43,8 @@
 #include "storage/primitive/empty_iterator.h"
 #include "storage/primitive/schema_helper.h"
 #include "storage/primitive/union_iterator.h"
+#include "storage/query/split_morsel_queue.h"
+#include "storage/query/split_scan_morsel.h"
 #include "storage/rowset/rowid_range_option.h"
 #include "storage/rowset/rowset_options.h"
 #include "storage/rowset/short_key_range_option.h"
