+++ be/src/storage/query/bucket_sequence_morsel_queue.h
@@ -16,8 +16,8 @@
 
 #include <mutex>
 
-#include "exec/pipeline/scan/olap_morsel_queue.h"
 #include "exec/pipeline/scan/ticketed_morsel_queue.h"
+#include "storage/query/olap_morsel_queue.h"
 
 namespace starrocks::pipeline {
