+++ be/src/storage/query/olap_dynamic_morsel_queue.h
@@ -19,8 +19,8 @@
 #include <mutex>
 #include <utility>
 
-#include "exec/pipeline/scan/olap_morsel_queue.h"
 #include "exec/pipeline/scan/ticketed_morsel_queue.h"
+#include "storage/query/olap_morsel_queue.h"
 
 namespace starrocks::pipeline {
