+++ be/src/storage/query/olap_morsel_queue.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "exec/pipeline/scan/olap_morsel_queue.h"
+#include "storage/query/olap_morsel_queue.h"
 
 namespace starrocks::pipeline {
