+++ be/src/storage/query/split_morsel_queue.h
@@ -16,11 +16,11 @@
 
 #include <mutex>
 
-#include "exec/pipeline/scan/olap_morsel_queue.h"
 #include "exec/pipeline/scan/ticketed_morsel_queue.h"
 #include "gutil/casts.h"
 #include "runtime/mem_pool.h"
 #include "storage/primitive/range.h"
+#include "storage/query/olap_morsel_queue.h"
 #include "storage/rowset/rowid_range_option.h"
 #include "storage/rowset/segment_group.h"
 #include "storage/rowset/short_key_range_option.h"
