+++ be/test/exec/connector_scan_node_test.cpp
@@ -24,13 +24,13 @@
 #include "compute_env/global_dict/fragment_dict_state.h"
 #include "exec/exec_factory.h"
 #include "exec/pipeline/scan/morsel.h"
-#include "exec/pipeline/scan/olap_morsel_queue.h"
 #include "gen_cpp/PlanNodes_types.h"
 #include "runtime/descriptor_helper.h"
 #include "runtime/exec_env.h"
 #include "runtime/runtime_state.h"
 #include "runtime/stream_load/load_stream_mgr.h"
 #include "runtime/stream_load/stream_load_pipe.h"
+#include "storage/query/olap_morsel_queue.h"
 
 namespace starrocks {
