+++ be/test/exec/pipeline/olap_scan_operator_test.cpp
@@ -17,12 +17,12 @@
 #include "compute_env/global_dict/fragment_dict_state.h"
 #include "exec/olap_scan_node.h"
 #include "exec/pipeline/query_context.h"
-#include "exec/pipeline/scan/olap_fixed_morsel_queue.h"
 #include "exec/pipeline/scan/olap_scan_prepare_operator.h"
 #include "gtest/gtest.h"
 #include "runtime/descriptors.h"
 #include "runtime/exec_env.h"
 #include "runtime/runtime_state.h"
+#include "storage/query/olap_fixed_morsel_queue.h"
 
 namespace starrocks::pipeline {
