+++ be/test/exec/pipeline/scan/morsel_queue_capability_test.cpp
@@ -16,14 +16,14 @@
 
 #include <memory>
 
-#include "exec/pipeline/scan/bucket_sequence_morsel_queue.h"
 #include "exec/pipeline/scan/dynamic_morsel_queue.h"
 #include "exec/pipeline/scan/fixed_morsel_queue.h"
-#include "exec/pipeline/scan/olap_dynamic_morsel_queue.h"
-#include "exec/pipeline/scan/olap_fixed_morsel_queue.h"
-#include "exec/pipeline/scan/olap_morsel_queue.h"
-#include "exec/pipeline/scan/split_morsel_queue.h"
 #include "exec/pipeline/scan/ticketed_morsel_queue.h"
+#include "storage/query/bucket_sequence_morsel_queue.h"
+#include "storage/query/olap_dynamic_morsel_queue.h"
+#include "storage/query/olap_fixed_morsel_queue.h"
+#include "storage/query/olap_morsel_queue.h"
+#include "storage/query/split_morsel_queue.h"
 
 namespace starrocks::pipeline {
