+++ be/test/platform/CMakeLists.txt
@@ -17,7 +17,6 @@ set(PLATFORM_TEST_FILES
     download_util_test.cpp
     key_cache_test.cpp
     llm/llm_cache_test.cpp
-    llm/llm_query_service_test.cpp
     llm/model_config_test.cpp
     path_rw_test.cpp
     platform_env_test.cpp
@@ -27,6 +26,17 @@ set(PLATFORM_TEST_FILES
     thrift_rpc_helper_test.cpp
     user_function_cache_test.cpp
 )
+
+if(NOT APPLE)
+    list(APPEND PLATFORM_TEST_FILES
+        llm/llm_query_service_test.cpp
+    )
+else()
+    # The HTTP-backed LLMQueryServiceTest cases pass on macOS but can hang in
+    # EvHttpServer::join() during suite teardown. Keep this suite in
+    # platform_test for non-Apple builds until the platform HTTP teardown issue
+    # is fixed.
+endif()
 
 set(PLATFORM_TEST_LINK_LIBS
     Platform
